Decompose common TS utility functions
