+++ core/static/core/display.css
@@ -38,7 +38,6 @@ td img {
     height: 2rem;
 }
 
-
 .stripey {
     background-color: var(--sweetwest-blue);
     background-image: linear-gradient(0deg,
