+++ core/templates/core/display.html
@@ -6,7 +6,7 @@
         <link rel="stylesheet" type=text/css href="{% static 'core/base.css' %}">
         <link rel="shortcut icon" type="image/svg" href="{% static 'core/favicon.svg' %}"/>
         <title>Sweetwest Airlines</title>
-        <link rel="stylesheet" type=text/css href="{% static 'core/table.css' %}">
+        <link rel="stylesheet" type=text/css href="{% static 'core/display.css' %}">
     </head>
     <body class="stripey">
         <table>
